style: move adapters and refactor httpserver class
