+++ src/project/util/gson/DurationAdapter.java
@@ -1,4 +1,4 @@
-package project.util;
+package project.util.gson;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
